Split up default reducers file
