+++ src/default-reducers/index.js
@@ -1,0 +1,5 @@
+export * from './create';
+export * from './read-one';
+export * from './read-many';
+export * from './update';
+export * from './delete';
