+++ src/default-reducers/update.js
@@ -1,0 +1,62 @@
+import {
+  updateResourcesMeta, updateManyResourcesMeta, upsertResource,
+  xhrStatuses, initialResourceMetaState
+} from '../utils';
+
+export function update(idAttr, state, action) {
+  const resourcesMeta = updateResourcesMeta(state.resourcesMeta, {
+    updatingStatus: xhrStatuses.PENDING
+  }, action[idAttr]);
+
+  return {
+    ...state,
+    resourcesMeta,
+  };
+}
+
+export function updateFailure(idAttr, state, action) {
+  const resourcesMeta = updateResourcesMeta(state.resourcesMeta, {
+    updatingStatus: xhrStatuses.FAILED
+  }, action[idAttr]);
+
+  return {
+    ...state,
+    resourcesMeta,
+  };
+}
+
+export function updateSuccess(idAttr, state, action) {
+  const resourcesMeta = updateResourcesMeta(state.resourcesMeta, {
+    updatingStatus: xhrStatuses.SUCCEEDED
+  }, action[idAttr]);
+
+  const resources = upsertResource(state.resources, action.resource, action[idAttr], idAttr);
+
+  return {
+    ...state,
+    resourcesMeta,
+    resources
+  };
+}
+
+export function updateAborted(idAttr, state, action) {
+  const resourcesMeta = updateResourcesMeta(state.resourcesMeta, {
+    updatingStatus: xhrStatuses.ABORTED
+  }, action[idAttr]);
+
+  return {
+    ...state,
+    resourcesMeta,
+  };
+}
+
+export function updateResetResolution(idAttr, state, action) {
+  const resourcesMeta = updateResourcesMeta(state.resourcesMeta, {
+    updatingStatus: xhrStatuses.NULL
+  }, action[idAttr]);
+
+  return {
+    ...state,
+    resourcesMeta,
+  };
+}
